Whitebox

Documentation for Whitebox.

Whitebox.default_callback

Whitebox.help

Whitebox.license

Whitebox.list_tools

Whitebox.run_tool

Whitebox.set_compress_rasters

Whitebox.set_verbose_mode

Whitebox.set_whitebox_dir

Whitebox.set_working_dir

Whitebox.to_camelcase

Whitebox.to_snakecase

Whitebox.tool_help

Whitebox.tool_parameters

Whitebox.toolbox

Whitebox.unzip

Whitebox.version

Whitebox.view_code

Whitebox.default_callback — Method

default_callback(value)

A simple default callback that outputs using the print function. When tools are called without providing a custom callback, this function will be used to print to standard output.

Whitebox.help — Method

help(wbt::WhiteboxTools)

Retrieves the help description for WhiteboxTools.

Whitebox.license — Function

license(wbt::WhiteboxTools, tool_name::Union{Nothing, String} = nothing)

Retrieves the license information for WhiteboxTools.

Whitebox.list_tools — Method

list_tools(wbt::WhiteboxTools, tool_name::String)

Lists all available tools in WhiteboxTools.

Whitebox.run_tool — Method

run_tool(wbt::WhiteboxTools, tool_name::String; callback=nothing, args::Union{Nothing, Vector{String}}=nothing)

Runs a tool and specifies tool arguments. Returns 0 if completes without error. Returns 1 on error with details.

Returns 2 if process is cancelled by user.

Whitebox.set_compress_rasters — Method

set_compress_rasters(wbt::WhiteboxTools, compress_rasters::Bool)

Sets the flag used by WhiteboxTools to determine whether to use compression for output rasters.

Whitebox.set_verbose_mode — Function

set_verbose_mode(wbt::WhiteboxTools, val::Bool=true)

Sets verbose mode. If verbose mode is False, tools will not print output messages. Tools will frequently provide substantial feedback while they are operating, e.g. updating progress for various sub-routines. When the user has scripted a workflow that ties many tools in sequence, this level of tool output can be problematic. By setting verbose mode to False, these

Whitebox.set_whitebox_dir — Method

set_whitebox_dir(path_str::String)

Sets the directory to the WhiteboxTools executable file.

Whitebox.set_working_dir — Method

set_working_dir(wbt::WhiteboxTools, path_str::String)

Sets the working directory, i.e. the directory in which the data files are located. By setting the working directory, tool input parameters that are files need only specify the file name rather than the complete file path.

Whitebox.to_camelcase — Method

to_camelcase(name::String)

Convert snake_case name to CamelCase name

Whitebox.to_snakecase — Method

to_snakecase(name::String)

Convert CamelCase name to snake_case name

Whitebox.tool_help — Method

tool_help(wbt::WhiteboxTools, tool_name::String)

Retrieves the help description for a specific tool.

Whitebox.tool_parameters — Method

tool_parameters(wbt::WhiteboxTools, tool_name::String)

Retrieves the tool parameter descriptions for a specific tool.

Whitebox.toolbox — Function

toolbox(wbt::WhiteboxTools, tool_name::String)

Retrieves the tool parameter descriptions for a specific tool.

Whitebox.unzip — Method

unzip(file::String; exdir::String="")

Unzip a .zip folder and put contents in exdir.

Modified from https://discourse.julialang.org/t/how-to-extract-a-file-in-a-zip-archive-without-using-os-specific-tools/34585

Whitebox.version — Method

version(wbt::WhiteboxTools)

Retrieves the version information for WhiteboxTools.

Whitebox.view_code — Method

view_code(wbt::WhiteboxTools, tool_name::String)

Opens a web browser to view the source code for a specific tool on the projects source code repository.